[Refactor] 🐾 type 파일 정의
