+++ src/mock/mockReservationsCall.ts
@@ -1,23 +1,10 @@
-export type ReservationSlots = {
-  [slot: string]: number;
-};
+import { ReservationCallDetail, ReservationSlots } from '@/types/reservation';
 
 export const mockReservations: ReservationSlots = {
   '09:00~10:00': 10,
   '10:00~11:00': 30,
   '11:00~12:00': 50,
   '12:00~13:00': 75,
-};
-
-export type ReservationCallDetail = {
-  id: string;
-  name: string;
-  phone: string;
-  consultationType: string;
-  date: string;
-  time: string;
-  waitingNumber: number;
-  waitingTime: number;
 };
 
 export const mockReservationCallDetails: ReservationCallDetail[] = [
